# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WordForge API
  description: |-
    WordForge 后端 API — v1-stable 端点。

    路径基础路径 `/api`（见 servers[0].url）。
    本规格由 `src/openapi.rs` 自动生成，通过 `cargo test --test openapi_export` 导出，CI 以 `git diff --exit-code docs/openapi.yaml` 防止规格漂移（drift）。
  version: 1.3.0-beta.1
servers:
- url: /api
paths:
...
  /resource-packs/public-key:
    get:
      tags:
      - resource-packs
      summary: 获取资源包签名 Ed25519 公钥
      description: 客户端硬编码公钥才是正式验签信任锚，本端点仅供 verify SDK 自检。`{publicKey, algorithm}` 裸 JSON。
      responses:
        '200':
          description: '`{publicKey: base64 string, algorithm: "ed25519"}`'
        '503':
          description: 签名器未初始化 — RESOURCE_PACK_SIGNER_UNAVAILABLE
  /resource-packs/{packId}/manifest:
    get:
      tags:
      - resource-packs
      summary: 获取指定 pack 当前 channel 激活版本的 manifest
      description: appVersion/locale 必填，channel 缺省 stable。支持 If-None-Match，命中返回 304。裸 JSON（非信封），字段名以 ResourcePackManifest 序列化输出为准（注意是 downloadURL，URL 三字母全大写）。
      parameters:
      - name: packId
        in: path
        required: true
        schema:
          type: string
      - name: appVersion
        in: query
        required: true
        schema:
          type: string
      - name: locale
        in: query
        required: true
        schema:
          type: string
      - name: channel
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: '`{packId, version, downloadURL, sha256, sizeBytes, minAppVersion, channel, signature, signatureAlgorithm}`'
        '304':
          description: If-None-Match 命中，无 body
        '400':
          description: 请求体校验失败 — VALIDATION_ERROR / INVALID_REQUEST_BODY
        '404':
          description: 资源包/该 channel 无激活版本 — RESOURCE_PACK_NOT_FOUND
        '409':
          description: appVersion 低于 minAppVersion — RESOURCE_PACK_APP_VERSION_TOO_LOW
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...